Modified a lot of scripts, and fixed the build error.
In summary:
Added a lot more code to the progression script, still yet to be tested. Enemies must be added first.
Modified the SceneMove script to migrate the handlers to the new scene correctly. This will need to be called on every single scene move to ensure handlers move correctly.
DontDestroyOnLoad was modified, however it may now be obslete. Testing without this is still required, this commit is a backup for the script.