Created a variant prefab of the enemy, updated progression.cs
The game now spawns enemies at random in the array. This means that you could have 5 enemies in the array and they'd have a set chance of spawning. I still need to modify this to ensure that all of the enemies are spawned, however I am not sure how I will do this.