UI Rework
UI has been reworked:
There are now dedicated buttons once the menu is paused. One button (Power Information) is currently disabled as it has not been programmed.
The player is now able to quit directly to the desktop by pressing some buttons.
The buttons are not currently working, will be fixed in the next build.