Updated player movement script
The player can now collide with enemies, I am going to update this to work with walls as well in the next push.
The next push will also include a modification to the enemy movement speed so it is possible to toggle the movement speed for enemies in the prefab's inspector tab.