Major UI rework
UI rework for the pause menu - Powers, the title and the descriptioons are all shown on the pause screen. Powers that are inactive have their alpha reduced.
Temporary icons have been added for these.
Meanwhile, in-game, I have replaced the text for modifiers to show all the powers icons instead.