Leaderboards name addition. UI bug fixes.
Changes:
Added the ability to add a name when you place on the leaderboard.
Moved the reset leaderboards and reset tutorial progress to the leaderboard window. There is currently no way to know if the leaderboards have been reset.
Fixed the leaderboards button to have clearer text.
Resetting leaderboards will now also reset the names for each slot.