Bug fixes, sword buffs
Changelog:
Fixed an issue when starting with the sword and trying to load the staff would cause the game to softlock.
Fixed an issue where when using the sword, effects such as justice rains from above and rain dance were 10x more likely to occur.
Fixed an issue where taking damage would not trigger the flash effect.
Increased the number of points on offer for losing health regeneration & double health powers.
The sword now has a permanent 25% damage reduction, which stacks with other powers such as Iron Maiden.
You are now able to use the sword while sprinting.