Lots of animation changes/additions
List of changes:
Added death animation to the micro orc, as well as the mega troll.
Fixed the death animations for the three base game enemies to no longer float upon death.
Added a spawning animation for the three base game enemies.
Added a roaring animation spawn for the final golem boss.