Small bug fixes
Changelog:
Reduced the timer for the first enemy spawn to 0.5 seconds, instead of having to wait an entire 7.5 seconds at the start.
Added a new “Loading…” when the player presses start on any of the difficulties.
Projectiles are now destroyed after 3.5 seconds if they are flying off into the distance, to prevent any anomalies.