AudioController code cleanup
Changelog:
Migrated PlayMagicAttack and PlayMeleeAttack into PlaySoundEffect as case 6 and 7 respectively.
Migrated all music tracks (excluding main menu, which is a separate instance to the rest of the game) to a single method, using IDs similar to PlaySoundEffect. 0 is for ingame music, 1 is for lose power music and 2 is for final boss music.