Bug fixes, code refactoring
Changelog:
Fixed all issues with migrating the tutorial code.
UIController: Removed TutorialStatus variable as it is now managed in the TutorialHandler.
UIController: Also removed all of the UI tutorial elements, as that is also now managed in the TutorialHandler.
UIController: Changed isPaused from public to a private variable. To access it outside of the class, use GetIsPaused(), and SetIsPaused().
Updated all occurrences of isPaused being accessed in public to use the new getter and setter accessors.
UIController: Reorganised variables on the top to have public, SerializeField and then private variables organised correctly.
Fixed some recurring errors where objects could not be found in loops due to the incorrect scene, so the object would never be able to be found.