Migration of Leaderboard to main menu.
Migrated leaderboard to now work on the main screen, instead of in it’s own scene. An animation has been added to move between each window.
Fixed a bug where the tag changes meant the player was no longer damaged.
Fixed a bug which meant that the melee enemy wouldn’t interact with the player correctly.
Increased the speed of the melee orc by 50%.