New credit for asset. New sounds. Bug fixes
Changelog:
Fixed bugs with the final boss.
Added new sounds with the final boss, with sounds for an enraged state.
Fixed some bugs with the UI, including fixing an issue where in expert mode the game would freeze due to the new timescaling, as well as moving the expert wave transition information to the middle of the screen.

Some sounds remain before completion of Regression!