Bug fixes, backup push before final testing session.
Bug fixes...:
Fixed an issue where the tutorial would not start correctly due to the new timing system.
Changed some text to be brighter and easier to read.
Updated all of the tutorial text, as well as the font, to make it easier to understand and read.
Updated hard/expert power removal to now tell the player what effect they have lost, to make it easier to understand.
Updated some power descriptions to be more descriptive to make it easier for the player to choose between.
Buffed mace damage by a further 50%. Mace is now the best weapon to use against the golem!
Slightly reduced the mega troll’s health. The final boss mega troll’s health remains the same.
Changed the stunned text to be far clearer to the user.
Changed the pause text to be entirely white to make it much easier to read.
Fixed a bug with the menus upon game completion to not load correctly.
Removed gloomy music when the game is completed.
Added a new message to show if the player has died or completed the game.
Fixed a rare issue where you could get multiple enemy progression and bonus points if an enemy was hit during their death animation.